requirements.txt
Pillow (pip install Pillow)
pillow_heif (pip install pillow pillow-heif)
